ble: add BLE support for the ESP32-S3
Split bt_ble.c into the shared driver plus per-chip layers behind
bt_ble.h (bt_ble_esp32c3.c, bt_ble_esp32s3.c).

The S3 cannot run the blob's ISRs in interrupt context because TinyGo dispatches
level-1 interrupts on the interrupted goroutine's stack and the
controller's call chain spills a register window per frame. This means the
handlers latch and mask, and the blob ISR runs from the scheduler tick.
Wiring uses CPU lines 13 and 17, the free level-1 lines. The S3 critical
section and phy_exit_critical() now restore only PS.INTLEVEL, not the
entire saved PS.

Needs the matching esp32s3 fixes in TinyGo in PR
https://github.com/tinygo-org/tinygo/pull/5600
