+++ ble.go
@@ -1,4 +1,4 @@
-//go:build esp32c3
+//go:build esp32c3 || esp32s3
 
 package espradio
 
@@ -10,22 +10,17 @@ import "C"
 import (
 	"errors"
 	"runtime"
-	"runtime/interrupt"
 	"time"
 	"unsafe"
 )
 
 // bleArenaPoolSize is the arena the BT controller allocates from.
 //
-// Note: the boot log shows the 10 ADV DATA TX buffers and 12 ACL TX buffers
-// coming back NULL with their EM_BASE_REG LUT entries set to 0x3fc00000, and the
-// DB heap getting 0 bytes. That looks like arena exhaustion but is not — raising
-// this to 96 KB reproduced the identical 22 NULLs, so those EM regions are
-// allocated lazily (on first advertise / first connection) and the 0x3fc00000
-// bases are simply "unmapped". Left at the WiFi size.
+// The 22 NULL TX buffers in the ESP32-C3 boot log are not arena exhaustion.
+// The controller allocates those EM areas at the first advertise or connection.
 const bleArenaPoolSize = arenaPoolSize
 
-// VHCITransport provides HCI read/write access to the ESP32 BLE controller
+// VHCITransport provides HCI read/write access to the BLE controller
 // via the Virtual HCI interface. It implements the transport interface
 // expected by the tinygo bluetooth package.
 type VHCITransport struct{}
@@ -69,7 +64,7 @@ func (t VHCITransport) Write(buf []byte) (int, error) {
 	return n, nil
 }
 
-// BLEInit initializes the BLE controller on the ESP32-C3.
+// BLEInit initializes the BLE controller.
 // Must be called before using VHCITransport.
 func BLEInit() error {
 	// Ensure radio hardware is powered up (clocks, power domain, modem reset).
@@ -109,36 +104,4 @@ func BLEInit() error {
 	time.Sleep(100 * time.Millisecond)
 
 	return nil
-}
-
-// ─── BT Interrupt Handlers ───────────────────────────────────────────────────
-
-// CPU interrupt numbers for BT (must match esp32c3/isr.c)
-const (
-	btCPUInt5 = 28 // RWBT + BT_BB → CPU interrupt 28
-	btCPUInt8 = 30 // RWBLE → CPU interrupt 30
-)
-
-func setupBTInterrupts() {
-	// Register and enable Go interrupt handlers (like WiFi does).
-	// Enable() sets priority=5 and edge-triggered, then
-	// espradio_bt_enable_hw_interrupts() overrides to priority=7 + level-triggered.
-	intr5 := interrupt.New(btCPUInt5, btISR5Handler)
-	intr5.Enable()
-	intr8 := interrupt.New(btCPUInt8, btISR8Handler)
-	intr8.Enable()
-}
-
-// Both dispatchers run blob code in real interrupt context, so they mark it:
-// anything they reach that would otherwise spin-and-yield must spin instead.
-func btISR5Handler(interrupt.Interrupt) {
-	C.espradio_enter_hw_isr()
-	C.espradio_bt_isr_dispatch_5()
-	C.espradio_exit_hw_isr()
-}
-
-func btISR8Handler(interrupt.Interrupt) {
-	C.espradio_enter_hw_isr()
-	C.espradio_bt_isr_dispatch_8()
-	C.espradio_exit_hw_isr()
 }
