+++ vhci_ring.go
@@ -1,4 +1,4 @@
-//go:build esp32c3 || esp32c3_qemu_target
+//go:build esp32c3 || esp32c3_qemu_target || esp32s3
 
 package espradio
